outlet/kafkaoutput: don't hard-code compression and load balancing
Same as for inlet, this should be configurable.
