+++ outlet/kafkaoutput/config.go
@@ -4,6 +4,8 @@
 package kafkaoutput
 
 import (
+	"github.com/twmb/franz-go/pkg/kgo"
+
 	"akvorado/common/kafka"
 )
 
@@ -14,11 +16,15 @@ type Configuration struct {
 	// deployments are unaffected.
 	Enabled             bool
 	kafka.Configuration `mapstructure:",squash" yaml:"-,inline"`
+	// CompressionCodec defines the compression to use.
+	CompressionCodec kafka.CompressionCodec
 	// QueueSize is the producer buffer: the max records held in flight
 	// (kgo MaxBufferedRecords) and the send-queue depth. When full, records are
 	// dropped, not blocked (best-effort; see dropped_messages_total and the
 	// kafka-output docs for sizing).
 	QueueSize int `validate:"min=1"`
+	// LoadBalance defines the load-balancing algorithm to use for Kafka partitions.
+	LoadBalance kafka.LoadBalanceAlgorithm
 }
 
 // DefaultConfiguration represents the default configuration for the Kafka output.
@@ -26,8 +32,10 @@ func DefaultConfiguration() Configuration {
 	cfg := kafka.DefaultConfiguration()
 	cfg.Topic = "flows-enriched"
 	return Configuration{
-		Enabled:       false,
-		Configuration: cfg,
-		QueueSize:     4096,
+		Enabled:          false,
+		Configuration:    cfg,
+		CompressionCodec: kafka.CompressionCodec(kgo.Lz4Compression()),
+		QueueSize:        4096,
+		LoadBalance:      kafka.LoadBalanceRandom,
 	}
 }
